refactor code to fix CI errors
