+++ care/facility/summarisation/triage_summary.py
@@ -1,26 +1,15 @@
 from celery.decorators import periodic_task
 from celery.schedules import crontab
 from django.db.models import Count, Sum
-from django.utils.decorators import method_decorator
 from django.utils.timezone import localtime, now
-from django.views.decorators.cache import cache_page
-from django_filters import rest_framework as filters
-from rest_framework.mixins import ListModelMixin
-from rest_framework.permissions import IsAuthenticatedOrReadOnly
-from rest_framework.viewsets import GenericViewSet
 
 from care.facility.models import Facility, FacilityPatientStatsHistory, FacilityRelatedSummary
-from care.facility.summarisation.facility_capacity import FacilitySummaryFilter, FacilitySummarySerializer
+from care.facility.summarisation.summary import SummaryViewSet
 
 
-class TriageSummaryViewSet(ListModelMixin, GenericViewSet):
-    lookup_field = "external_id"
-    queryset = FacilityRelatedSummary.objects.filter(s_type="TriageSummary").order_by("-created_date")
-    permission_classes = (IsAuthenticatedOrReadOnly,)
-    serializer_class = FacilitySummarySerializer
-
-    filter_backends = (filters.DjangoFilterBackend,)
-    filterset_class = FacilitySummaryFilter
+class TriageSummaryViewSet(SummaryViewSet):
+    def get_queryset(self):
+        return super().get_queryset().filter(s_type="TriageSummary")
 
     # def get_queryset(self):
     #     user = self.request.user
@@ -32,10 +21,6 @@ class TriageSummaryViewSet(ListModelMixin, GenericViewSet):
     #     elif self.request.user.user_type >= User.TYPE_VALUE_MAP["StateReadOnlyAdmin"]:
     #         return queryset.filter(facility__state=user.state)
     #     return queryset.filter(facility__users__id__exact=user.id)
-
-    @method_decorator(cache_page(60 * 60 * 1))
-    def dispatch(self, request, *args, **kwargs):
-        return super().dispatch(request, *args, **kwargs)
 
 
 def TriageSummary():
